ocr cleanup
