+++ paperlessOCR/src/main/java/org/openapitools/paperlessocr/persistence/elasticsearch/entities/ElasticDocumentDocument.java
@@ -18,6 +18,4 @@ public class ElasticDocumentDocument {
     private String title;
 
     private String content;
-
-    private String filename;
 }
